+++ users/views.py
@@ -2,20 +2,30 @@ from django.shortcuts import render
 
 # Create your views here.
 
-from rest_framework.views import APIView
+from rest_framework import generics, status
 from rest_framework.response import Response
 from rest_framework import status
 from .serializers import RegisterSerializer
+from rest_framework.permissions import AllowAny
 
-class RegisterView(APIView):
-    permission_classes = []
 
-    def post(self, request):
-        serializer = RegisterSerializer(data=request.data)
+class RegisterView(generics.CreateAPIView):
+    serializer_class = RegisterSerializer
+    permission_classes = [AllowAny]
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
-        serializer.save()
+        user = serializer.save()
 
         return Response(
-            {"message": "User registered successfully"},
-            status=status.HTTP_201_CREATED
+            {
+                "user": {
+                    "id": user.id,
+                    "email": user.email,
+                    "name": user.name,
+                },
+                "message": "User registered successfully.",
+            },
+            status=status.HTTP_201_CREATED,
         )