+++ wallets/models/transaction.py
@@ -1,0 +1,47 @@
+from django.conf import settings
+from django.db import models
+from django.utils import timezone
+
+from .wallet import Wallet
+
+class Transaction(models.Model):
+    class Type(models.TextChoices):
+        DEPOSIT = "deposit", "Deposit"
+        WITHDRAW = "withdraw", "Withdraw"
+        TRANSFER_IN = "transfer_in", "Transfer In"
+        TRANSFER_OUT = "transfer_out", "Transfer Out"
+    
+    wallet = models.ForeignKey(
+        Wallet,
+        on_delete=models.PROTECT,
+        related_name="transactions"
+    )
+    user = models.ForeignKey(
+    settings.AUTH_USER_MODEL,
+    on_delete=models.PROTECT,
+    related_name="transactions",
+)
+
+    amount = models.DecimalField(
+        max_digits=12,
+        decimal_places=2
+    )
+
+    type = models.CharField(
+        max_length=20,
+        choices=Type.choices
+    )
+
+    reference = models.CharField(
+        max_length=100,
+        blank=True,
+        null=True
+    )
+
+    created_at = models.DateTimeField(default=timezone.now)
+
+    class Meta:
+        ordering = ["-created_at"]
+
+    def __str__(self):
+        return f"{self.type} - {self.amount}"